requirements.txt
Django
daphne
channels
channels-redis
djangorestframework
drf-spectacular
pymongo
redis
dnspython